requirements.txt
streamlit>=1.28.0
paho-mqtt>=1.6.0
